# ironflow-sdk/openapi.json
{
  "components": {
    "schemas": {
...
      "AuditLogEntry": {
        "description": "A persisted audit log entry capturing a domain event.\n\nEach entry records the full event payload as JSON alongside\ndenormalized contextual IDs for efficient filtering.\n\n# Examples\n\n```\nuse ironflow_store::entities::{AuditLogEntry, EventKind};\nuse uuid::Uuid;\nuse chrono::Utc;\nuse serde_json::json;\n\nlet entry = AuditLogEntry {\n    id: Uuid::now_v7(),\n    event_type: EventKind::RunStatusChanged,\n    payload: json!({\"run_id\": \"...\"}),\n    run_id: Some(Uuid::now_v7()),\n    step_id: None,\n    user_id: None,\n    created_at: Utc::now(),\n};\nassert_eq!(entry.event_type, EventKind::RunStatusChanged);\n```",
        "properties": {
          "created_at": {
            "description": "When the event was recorded.",
            "format": "date-time",
            "type": "string"
          },
          "event_type": {
            "$ref": "#/components/schemas/EventKind",
            "description": "Event type."
          },
          "id": {
            "description": "Unique entry ID (UUID v7).",
            "format": "uuid",
            "type": "string"
          },
          "payload": {
            "description": "Full serialized event payload."
          },
          "run_id": {
            "description": "Associated run ID, if the event is run-scoped.",
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "step_id": {
            "description": "Associated step ID, if the event is step-scoped.",
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "user_id": {
            "description": "Associated user ID, if the event is user-scoped.",
            "format": "uuid",
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "id",
          "event_type",
          "payload",
          "created_at"
        ],
        "type": "object"
      },
...
      "CreateRunRequest": {
        "description": "Request to trigger a workflow.\n\n# Examples\n\n```\nuse ironflow_api::entities::CreateRunRequest;\nuse serde_json::json;\n\nlet req = CreateRunRequest {\n    workflow: \"deploy\".to_string(),\n    payload: Some(json!({\"env\": \"prod\"})),\n    labels: None,\n    scheduled_at: None,\n    max_retries: Some(2),\n    max_cost_usd: None,\n};\nassert_eq!(req.workflow, \"deploy\");\n```",
        "properties": {
          "labels": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "Optional key-value labels for categorization and filtering.",
            "nullable": true,
            "propertyNames": {
              "type": "string"
            },
            "type": "object"
          },
          "max_cost_usd": {
            "description": "Optional cumulative cost cap for this run, in USD.\n\nOverrides the workflow default and the server default. `None` falls back\nto those. Must be zero or positive.",
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "max_retries": {
            "description": "How many times the run may be replayed automatically after a transient\nfailure. Defaults to `0`, meaning no automatic retry.\n\nEach retry waits an exponential backoff (30 s, 2 min, 8 min, capped at\n15 min) before the run is replayed from the start. Failures that cannot\nsucceed on replay -- an unknown workflow, an invalid payload, an\nexhausted agent budget, a rejected approval, a manual cancellation --\nconsume no attempt.",
            "format": "int32",
            "minimum": 0,
            "nullable": true,
            "type": "integer"
          },
          "payload": {
            "additionalProperties": {},
            "description": "Optional input payload for the workflow.",
            "nullable": true,
            "propertyNames": {
              "type": "string"
            },
            "type": "object"
          },
          "scheduled_at": {
            "description": "Optional deferred execution time. `None` means run immediately.",
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "workflow": {
            "description": "The workflow name to trigger.",
            "type": "string"
          }
        },
        "required": [
          "workflow"
        ],
        "type": "object"
      },
...
      "EventKind": {
        "description": "Strongly-typed event kind matching domain event variants.\n\nSerializes to/from `snake_case` strings (e.g. `\"run_status_changed\"`).\n\n# Examples\n\n```\nuse ironflow_store::entities::EventKind;\n\nlet kind: EventKind = \"run_status_changed\".parse().unwrap();\nassert_eq!(kind, EventKind::RunStatusChanged);\nassert_eq!(kind.as_str(), \"run_status_changed\");\n```",
        "enum": [
          "run_created",
          "run_status_changed",
          "run_failed",
          "run_budget_exceeded",
          "step_completed",
          "step_failed",
          "approval_requested",
          "approval_granted",
          "approval_rejected",
          "approval_escalated",
          "log_line",
          "user_signed_in",
          "user_signed_up",
          "user_signed_out",
          "secrets_rotated",
          "retry_forced",
          "provider_account.updated",
          "provider_account.usage_updated"
        ],
        "type": "string"
      },
...
      "RunStatus": {
        "description": "Status of a workflow run, forming a finite state machine.\n\nValid transitions:\n- `Pending` -> `Running`, `Cancelled`\n- `Running` -> `Pending` (worker lease expired), `Completed`, `Failed`, `Warning`, `Retrying`, `Cancelled`, `AwaitingApproval`, `Sleeping`\n- `Retrying` -> `Running`, `Failed`, `Cancelled`\n- `AwaitingApproval` -> `Running`, `Pending` (requeued for a worker under `ExecutionMode::Workers`), `Failed`, `Cancelled`\n- `Sleeping` -> `Pending` (wake-up timer elapsed), `Cancelled`\n\nTerminal states (`Completed`, `Failed`, `Warning`, `Cancelled`) are idempotent:\ntransitioning to the same terminal state is a no-op, not an error.\n\n# Examples\n\n```\nuse ironflow_store::entities::RunStatus;\n\nassert!(RunStatus::Pending.can_transition_to(&RunStatus::Running));\nassert!(!RunStatus::Pending.can_transition_to(&RunStatus::Completed));\nassert!(!RunStatus::Completed.can_transition_to(&RunStatus::Running));\nassert!(RunStatus::Running.can_transition_to(&RunStatus::AwaitingApproval));\nassert!(RunStatus::Running.can_transition_to(&RunStatus::Warning));\n// A run whose worker lease expired goes back to the queue:\nassert!(RunStatus::Running.can_transition_to(&RunStatus::Pending));\nassert!(RunStatus::AwaitingApproval.can_transition_to(&RunStatus::Running));\nassert!(RunStatus::AwaitingApproval.can_transition_to(&RunStatus::Pending));\nassert!(RunStatus::Running.can_transition_to(&RunStatus::Sleeping));\nassert!(RunStatus::Sleeping.can_transition_to(&RunStatus::Pending));\nassert!(RunStatus::Sleeping.can_transition_to(&RunStatus::Cancelled));\n// Terminal-to-same is idempotent:\nassert!(RunStatus::Failed.can_transition_to(&RunStatus::Failed));\nassert!(RunStatus::Completed.can_transition_to(&RunStatus::Completed));\nassert!(RunStatus::Warning.can_transition_to(&RunStatus::Warning));\nassert!(RunStatus::Cancelled.can_transition_to(&RunStatus::Cancelled));\n```",
        "enum": [
          "pending",
          "running",
          "completed",
          "failed",
          "retrying",
          "cancelled",
          "awaiting_approval",
          "warning",
          "sleeping"
        ],
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Thomas Tartrau"
    },
    "description": "REST API for the ironflow workflow engine",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "Ironflow REST API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/audit-logs": {
      "get": {
        "description": "Admin-only. Returns a paginated list of persisted domain events\nfor compliance review and post-mortem debugging.",
        "operationId": "list_audit_logs",
        "parameters": [
          {
            "description": "Filter by event type (e.g. `run_status_changed`).",
            "in": "query",
            "name": "event_type",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/EventKind"
            }
          },
          {
            "description": "Filter by run ID.",
            "in": "query",
            "name": "run_id",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "Filter entries created at or after this timestamp.",
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "description": "Filter entries created at or before this timestamp.",
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "description": "Page number (1-based, default: 1).",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "description": "Items per page (default: 50, max: 100).",
            "in": "query",
            "name": "per_page",
            "required": false,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditLogEntry"
                  },
                  "type": "array"
                }
              }
            },
            "description": "List of audit log entries with pagination"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - admin only"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "List audit log entries with optional filtering and pagination.",
        "tags": [
          "audit"
        ]
      }
    },
...
    "/api/v1/runs": {
      "get": {
        "description": "# Query Parameters\n\n- `workflow` — Filter by workflow name (optional)\n- `status` — Filter by run status (optional)\n- `created_by` — Filter by author user ID (optional). Also matches runs\n  triggered by one of that user's API keys.\n- `page` — Page number, 1-based (default: 1)\n- `per_page` — Items per page (default: 20, max: 100)",
        "operationId": "list_runs",
        "parameters": [
          {
            "description": "Filter by workflow name.",
            "in": "query",
            "name": "workflow",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter by run status.",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/RunStatus"
            }
          },
          {
            "description": "Filter by step presence (only applies to completed/cancelled runs).\nNon-terminal runs (pending, running, etc.) are always included.\nWhen `true`, only return completed/cancelled runs that have steps.\nWhen `false`, only return completed/cancelled runs without steps.",
            "in": "query",
            "name": "has_steps",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "description": "Filter by labels. Comma-separated `key:value` pairs.",
            "in": "query",
            "name": "label",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter by author: the user ID that triggered the run.\n\nAlso matches runs triggered by one of that user's API keys.",
            "in": "query",
            "name": "created_by",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "Page number (1-based).",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "description": "Items per page.",
            "in": "query",
            "name": "per_page",
            "required": false,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "List of runs with pagination"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "List runs with optional filtering and pagination.",
        "tags": [
          "runs"
        ]
      },
      "post": {
        "description": "Returns 201 Created with the newly enqueued run.\n\nAn optional `Idempotency-Key` header makes the call safe to replay: the same\nkey returns the run it already produced with 200 OK instead of enqueueing a\nsecond one. A key reused with a different workflow or payload is rejected with\n409 Conflict. Keys stay bound for 24 hours, after which they are released.\n\n# Errors\n\nReturns [`ApiError::Forbidden`] for non-admin callers.\nReturns [`ApiError::BadRequest`] if the workflow is unknown, the body is\ninvalid, or the `Idempotency-Key` header is malformed.\nReturns [`ApiError::IdempotencyKeyConflict`] if the key is bound to a\ndifferent request.\nReturns [`ApiError::MonthlyBudgetExceeded`] if the global monthly cost quota\nis exhausted.",
        "operationId": "create_run",
        "parameters": [
          {
            "description": "Optional key making the call safe to replay. At most 255 printable ASCII characters, valid for 24 hours.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          },
          "description": "Workflow to trigger",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Idempotency key replayed: the existing run is returned"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Run created successfully"
          },
          "400": {
            "description": "Unknown workflow, invalid body or malformed Idempotency-Key"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Idempotency key already used with a different request"
          },
          "429": {
            "description": "Monthly cost quota exhausted"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Trigger a workflow by name.",
        "tags": [
          "runs"
        ]
      }
    },
...
  },
...
}